WW-4996 Refactors creator to use ObjectFactory#buildConverter method Moves class to a proper package and renames it to better expose its function as a default Struts implementation
