+++ core/src/main/java/org/apache/struts2/conversion/StrutsTypeConverterCreator.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.conversion.impl;
+package org.apache.struts2.conversion;
 
 import com.opensymphony.xwork2.ObjectFactory;
 import com.opensymphony.xwork2.conversion.TypeConverter;
@@ -27,7 +27,7 @@ import com.opensymphony.xwork2.ognl.XWorkTypeConverterWrapper;
 /**
  * Default implementation of {@link TypeConverterCreator}
  */
-public class DefaultTypeConverterCreator implements TypeConverterCreator {
+public class StrutsTypeConverterCreator implements TypeConverterCreator {
 
     private ObjectFactory objectFactory;
 
@@ -37,24 +37,22 @@ public class DefaultTypeConverterCreator implements TypeConverterCreator {
     }
 
     public TypeConverter createTypeConverter(String className) throws Exception {
-        Object obj = objectFactory.buildBean(className, null);
-        return getTypeConverter(obj);
+        Class<?> classInstance = objectFactory.getClassInstance(className);
+
+        if (classInstance.isAssignableFrom(TypeConverter.class)) {
+            Class<? extends TypeConverter> converterClass = (Class<? extends TypeConverter>) classInstance;
+            return objectFactory.buildConverter(converterClass, null);
+        } else {
+            throw new IllegalArgumentException("Type converter class " + className + " doesn't implement " + TypeConverter.class.getName());
+        }
     }
 
     public TypeConverter createTypeConverter(Class<?> clazz) throws Exception {
-        Object obj = objectFactory.buildBean(clazz, null);
-        return getTypeConverter(obj);
-    }
-
-    protected TypeConverter getTypeConverter(Object obj) {
-        if (obj instanceof TypeConverter) {
-            return (TypeConverter) obj;
-
-            // For backwards compatibility
-        } else if (obj instanceof ognl.TypeConverter) {
-            return new XWorkTypeConverterWrapper((ognl.TypeConverter) obj);
+        if (clazz.isAssignableFrom(TypeConverter.class)) {
+            Class<? extends TypeConverter> converterClass = (Class<? extends TypeConverter>) clazz;
+            return objectFactory.buildConverter(converterClass, null);
         } else {
-            throw new IllegalArgumentException("Type converter class " + obj.getClass() + " doesn't implement com.opensymphony.xwork2.conversion.TypeConverter");
+            throw new IllegalArgumentException("Type converter class " + clazz.getName() + " doesn't implement " + TypeConverter.class.getName());
         }
     }
